+++ CHANGELOG.md
@@ -6,11 +6,10 @@
 
 #### Changed
 
-- [#972](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/972) Add support to `ActiveRecord::QueryLogs`
 ...
 
 #### Added
 
-...
+- [#972](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/972) Add support to `ActiveRecord::QueryLogs`
 
 Please check [6-1-stable](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/blob/6-1-stable/CHANGELOG.md) for previous changes.